Add t.before(fn) method (#28)
While the recommendation in the past has been 'use t.test() for this',
it's really annoying to have to specify `-g 'setup|my test'` when
running a single test for debugging.  Also, you might not want to have
assertions for these; often just a simple pass/fail is enough, and can
keep the test output a bit quieter.

`t.before(fn)` is effectively a single (optionally promise-returning)
function which can be run before all tests and assertions, and which is
immune to any filters.

If the `t.before(fn)` function throws or rejects, is called more than
once, or is called after any tests or assertions, then the test fails.