separate t.match() and t.has()
t.has() is a way to test for the fields specified, _without_ fancy
regexps and type matching.

t.match() is a way to test for regexp matches, types, and lots of other
conveniences.

With this change, there's (finally) a direct 1:1 correlation between the
methods exported by tcompare and the associated tap assertion functions.