Improve tap-snapshot folder structure
Previously, every test would save its snapshots in a file in
`tap-snapshots` with a name that was the sanitized `t.fullname`.  And
since each test uses the snapshot file from its parent test, the
test.name was almost always 'TAP'.

This led to a lot of files named things like
`tap-snapshots/test-lib-utils-base-utils.js-TAP.test.cjs`.  When
snapshots had to be examined or reasoned about, it made it unnecessarily
difficult.

This changes the naming convention to instead take the relative path
from the current working directory to the main script as-is, and root it
on the `tap-snapshots` folder.  Then, any arguments beyond the script
name are sanitized and appended, and finally the `.test.cjs` extension
is added.

Thus, for the example above, you get
`tap-snapshots/test/lib/utils/base.js.test.cjs` as the snapshot name for
`test/lib/utils/base.js`, which makes the folder _much_ more covenient
to work with on larger projects, and less repetitive in the filenames.