Make the snapshot file location fully customizable (#38)
Overriding settings.snapshotFile lets you put anything you want in
there.  Of course, using this badly will result in snapshots not
working, but that comes with the territory.  This makes it quite easy
for users to store snapshots for a file at `test/routes/foo.js` in
`test/routes/foo.snap`, or anywhere else they like.