Improve error messages reported by t.resolves()
When the promise passed to `t.resolves()` is rejected:

Before this change, the test would fail with a generic error message
 ✖ expect resolving Promise

This is rather unhelpful, as the underlying error with details about
the actual cause is hidden from the user. Also the error location
is pointing to `t.resolves()`.

```
  2 |
  3 | test('resolve-reject', async (t) => {
> 4 |   await t.resolves(Promise.reject(new Error('rejection')));
    | ----^
  5 | });
```

In this commit, I have reworked the implementation of `t.resolves()`
to use the same solution as in `t.doesNotThrow()` and call `t.fail()`
when the Promise was rejected.

As a result, information about the original error is lifted up
and the error location is pointing to the place where the original error
was created.

```
  2 |
  3 | test('resolve-reject', async (t) => {
> 4 |   await t.resolves(Promise.reject(new Error('rejection')));
    | ----------------------------^
  5 | });
```

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>