subtests: only output comment style
Consensus on TAP 14 specification is that 'buffered subtests'
will not be in the spec.

It's no harder to emit the comment style every time.  The downside
is purely ergonomic, since it becomes impossible to know whether
the subtest was run in parallel or not, but that's not relevant
from the harness's point of view anyway, in most cases.

See: https://github.com/TestAnything/Specification/issues/29